backend/requirements.txt
# Django Core
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.3

# Auth
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0

# Database
psycopg[binary]>=3.1

# CORS
django-cors-headers==4.4.0

# Environment
python-decouple==3.8

# Async Tasks (optional)
# celery==5.4.0
# redis==5.0.8

# Dev Tools
django-extensions==3.2.3
drf-spectacular==0.27.2  # OpenAPI/Swagger
